% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cb_2018_il}
\alias{cb_2018_il}
\title{Information for the results of the presidental election in Turkey by province in 2018}
\format{
A data frame with 81 rows and 105 variables:
\describe{
  \item{Il_Adi}{il adi, the name of the province}
  \item{Kayitli_Secmen_Sayisi}{kayitli secmen sayisi, the number of voters}
  \item{Oy_Kullanan_Secmen_Sayisi}{oy kullanan secmen sayisi, the number of voters who vote}
  \item{Gecerli_Oy_Toplami}{gecerli oy sayisi, the number of valid votes }
  \item{MUHARREM_INCE}{muharrem incenin aldigi oy sayisi, the number of valid votes that muharrem ince received }
  \item{MERAL_AKSENER}{meral aksenerin aldigi oy sayisi, the number of valid votes that meral aksener received }
  \item{RECEP_TAYYIP_ERDOGAN}{recep tayyip erdoganin aldigi oy sayisi, the number of valid votes that recep tayyip erdogan received }
  \item{SELAHATTIN_DEMIRTAS}{selahattin demirtasin aldigi oy sayisi, the number of valid votes that selahattin demirtas received}
  \item{TEMEL_KARAMOLLAOGLU}{temel karamollaoglunun aldigi oy sayisi, the number of valid votes that temel karamollaoglu received }
  \item{DOGU_PERINCEK}{dogu perincekin aldigi oy sayisi, the number of valid votes that dogu perincek received }
  \item{MUHARREM_INCE_\%}{muharrem incenin aldigi oy yuzdesi, the percentage of valid votes that muharrem ince received }
  \item{MERAL_AKSENER_\%}{meral aksenerin aldigi oy yuzdesi, the percentage of valid votes that meral aksener received }
  \item{RECEP_TAYYIP_ERDOGAN_\%}{recep tayyip erdoganin aldigi oy yuzdesi, the percentage of valid votes that recep tayyip erdogan received }
  \item{SELAHATTIN_DEMIRTAS_\%}{selahattin demirtasin aldigi oy yuzdesi, the percentage of valid votes that selahattin demirtas received}
  \item{TEMEL_KARAMOLLAOGLU_\%}{temel karamollaoglunun aldigi oy yuzdesi, the percentage of valid votes that temel karamollaoglu received }
  \item{DOGU_PERINCEK_\%}{dogu perincekin aldigi oy yuzdesi, the percentage of valid votes that dogu perincek received }
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
cb_2018_il
}
\description{
A dataset showing the information for the candidates competing in the presidental election, and their votes in count and percentage.
}
\keyword{datasets}
