% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cb_2018_turkiye}
\alias{cb_2018_turkiye}
\title{Information for the results of the presidental election in Turkey in 2018.}
\format{
A data frame with 6 rows and 3 variables:
\describe{
  \item{Aday}{aday adi, name of the candidate}
  \item{Oy}{alinan oy sayisi, the number of votes received}
  \item{Yuzde}{alinan oy sayisi (yuzde olarak), the percentage of votes received}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
cb_2018_turkiye
}
\description{
A dataset showing the information for the candidates competing in the presidental election, and their votes in count and percentage.
}
\keyword{datasets}
