% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{cb_2023_ilk_tur_il}
\alias{cb_2023_ilk_tur_il}
\title{Information for the results of the first round of the presidential election in Turkey by province in 2023}
\format{
A data frame with 81 rows and 12 variables:
\describe{
\item{Il_Adi}{il adi, the name of the province}
\item{Kayitli_Secmen_Sayisi}{kayitli secmen sayisi, the number of registered voters}
\item{Oy_Kullanan_Secmen_Sayisi}{oy kullanan secmen sayisi, the number of voters who vote}
\item{Gecerli_Oy_Toplami}{gecerli oy toplami, the total number of valid votes}
\item{RECEP_TAYYIP_ERDOGAN}{recep tayyip erdoganin aldigi oy sayisi, the number of valid votes that Recep Tayyip Erdogan received}
\item{MUHARREM_INCE}{muharrem incenin aldigi oy sayisi, the number of valid votes that Muharrem Ince received}
\item{KEMAL_KILICDAROGLU}{kemal kilicdaroglunun aldigi oy sayisi, the number of valid votes that Kemal Kilicdaroglu received}
\item{SINAN_OGAN}{sinan oganin aldigi oy sayisi, the number of valid votes that Sinan Ogan received}
\item{RECEP_TAYYIP_ERDOGAN_\%}{recep tayyip erdoganin aldigi oy yuzdesi, the percentage of valid votes that Recep Tayyip Erdogan received}
\item{MUHARREM_INCE_\%}{muharrem incenin aldigi oy yuzdesi, the percentage of valid votes that Muharrem Ince received}
\item{KEMAL_KILICDAROGLU_\%}{kemal kilicdaroglunun aldigi oy yuzdesi, the percentage of valid votes that Kemal Kilicdaroglu received}
\item{SINAN_OGAN_\%}{sinan oganin aldigi oy yuzdesi, the percentage of valid votes that Sinan Ogan received}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
cb_2023_ilk_tur_il
}
\description{
A dataset showing the information for the candidates competing in the first round of the presidential election, and their votes in count and percentage.
}
\keyword{datasets}
