% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_list.R
\name{data_list}
\alias{data_list}
\title{data_list}
\usage{
data_list()
}
\value{
A data frame representing the file names and election names
}
\description{
list the name of the data sets with the election names
}
\examples{
data_list()
}
