% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{genel_secim_2002_turkiye}
\alias{genel_secim_2002_turkiye}
\title{Information for the results of the general election in Turkey in 2002.}
\format{
A data frame with 19 rows and 5 variables:
\describe{
  \item{Parti}{parti adi, name of the political party}
  \item{Genel_Baskan}{parti baskani adi, name of the head of the party}
  \item{Oy}{alinan oy sayisi, the number of votes received}
  \item{Yuzde}{alinan oy sayisi (yuzde olarak), the percentage of votes received}
  \item{Sandalye}{secim sonucu kazanilan sandalye sayisi, the number of seats obtained in the council after the election}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
genel_secim_2002_turkiye
}
\description{
A dataset showing the information for party competing in the election, head of party, vote in count and percentage and the number of seats obtained in the council after the election region in 2002 general election in Turkey.
}
\keyword{datasets}
