% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{genel_secim_2018_il}
\alias{genel_secim_2018_il}
\title{Information for the results of the general election in Turkey by province in 2018}
\format{
A data frame with 81 rows and 12 variables:
\describe{
\item{Il}{il, the name of the province}
\item{MV}{mv, the number of parliamentary seats in each province}
\item{AK_Parti}{ak parti oy sayisi, the number of votes that the Justice and Development Party (AK Parti) received}
\item{AK_Parti_\%}{ak parti oy yuzdesi, the percentage of votes that the Justice and Development Party (AK Parti) received}
\item{CHP}{chp oy sayisi, the number of votes that the Republican People's Party (CHP) received}
\item{CHP_\%}{chp oy yuzdesi, the percentage of votes that the Republican People's Party (CHP) received}
\item{MHP}{mhp oy sayisi, the number of votes that the Nationalist Movement Party (MHP) received}
\item{MHP_\%}{mhp oy yuzdesi, the percentage of votes that the Nationalist Movement Party (MHP) received}
\item{HDP}{hdp oy sayisi, the number of votes that the Peoples' Democratic Party (HDP) received}
\item{HDP_\%}{hdp oy yuzdesi, the percentage of votes that the Peoples' Democratic Party (HDP) received}
\item{IYI_Parti}{iyi parti oy sayisi, the number of votes that the Good Party (IYI Parti) received}
\item{IYI_Parti_\%}{iyi parti oy yuzdesi, the percentage of votes that the Good Party (IYI Parti) received}
\item{Toplam}{toplam oy sayisi, the total number of votes cast in each province}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
genel_secim_2018_il
}
\description{
A dataset showing the information for the political parties competing in the general election, their vote counts, percentages, and the number of parliamentary seats by province.
}
\keyword{datasets}
