% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{genel_secim_2023_il}
\alias{genel_secim_2023_il}
\title{Information for the results of the general election in Turkey by province in 2023}
\format{
A data frame with 81 rows and 63 variables:
\describe{
  \item{Il}{il, the name of the province}
  \item{Kayitli_Secmen_Sayisi}{kayitli secmen sayisi, the number of registered voters}
  \item{Oy_Kullanan_Secmen_Sayisi}{oy kullanan secmen sayisi, the number of voters who vote}
  \item{Gecerli_Oy_Toplami}{gecerli oy toplami, the total number of valid votes}
  \item{MILLET}{millet partisinin oy sayisi, the number of votes the Nation Party received}
  \item{HAK-PAR}{hak-par'in oy sayisi, the number of votes the Rights and Freedoms Party received}
  \item{TKP}{tkp'nin oy sayisi, the number of votes the Communist Party of Turkey received}
  \item{TKH}{tkh'nin oy sayisi, the number of votes the People's Liberation Party received}
  \item{SOL_PARTI}{sol parti'nin oy sayisi, the number of votes the Left Party received}
  \item{GENCPARTI}{gencparti'nin oy sayisi, the number of votes the Young Party received}
  \item{MEMLEKET}{memleket partisinin oy sayisi, the number of votes the Homeland Party received}
  \item{BBP}{bbp'nin oy sayisi, the number of votes the Great Unity Party received}
  \item{AK_Parti}{ak parti'nin oy sayisi, the number of votes the Justice and Development Party received}
  \item{YENIDEN_REFAH}{yeniden refah partisinin oy sayisi, the number of votes the Refah Party received}
  \item{MHP}{mhp'nin oy sayisi, the number of votes the Nationalist Movement Party received}
  \item{YESIL_SOL_PARTI}{yesil sol partinin oy sayisi, the number of votes the Green Left Party received}
  \item{TIP}{tip'in oy sayisi, the number of votes the Labour Party received}
  \item{AB}{ab'nin oy sayisi, the number of votes the Bloc Party received}
  \item{ANAP}{anap'in oy sayisi, the number of votes the Motherland Party received}
  \item{YP}{yp'nin oy sayisi, the number of votes the New Path Party received}
  \item{HKP}{hkp'nin oy sayisi, the number of votes the People's Liberation Party received}
  \item{MILLI_YOL}{milli yol'un oy sayisi, the number of votes the National Course Party received}
  \item{VATAN_PARTISI}{vatan partisinin oy sayisi, the number of votes the Homeland Party received}
  \item{GBP}{gbp'nin oy sayisi, the number of votes the Independent Turkey Party received}
  \item{CHP}{chp'nin oy sayisi, the number of votes the Republican People's Party received}
  \item{IYI_Parti}{iyi parti'nin oy sayisi, the number of votes the Good Party received}
  \item{AP}{ap'nin oy sayisi, the number of votes the Alternative Party received}
  \item{ZAFER_PARTISI}{zafer partisinin oy sayisi, the number of votes the Victory Party received}
  \item{SOSYALIST_GUC_BIRLIGI_ITTIFAKI}{sosyalist guc birligi ittifakinin oy sayisi, the number of votes the Socialist Power Union Alliance received}
  \item{CUMHUR_ITTIFAKI}{cumhur ittifakinin oy sayisi, the number of votes the People's Alliance received}
  \item{EMEK_VE_OZGURLUK_ITTIFAKI}{emek ve ozgurluk ittifakinin oy sayisi, the number of votes the Labour and Freedom Alliance received}
  \item{MILLET_ITTIFAKI}{millet ittifakinin oy sayisi, the number of votes the Nation Alliance received}
  \item{ATA_ITTIFAKI}{ata ittifakinin oy sayisi, the number of votes the Ancestor Alliance received}
  \item{BAGIMSIZ_TOPLAM_OY}{bagimsiz adaylarin toplam oy sayisi, the total number of votes independent candidates received}
  \item{MILLET_\%}{millet partisinin oy yuzdesi, the percentage of votes the Nation Party received}
  \item{HAK-PAR_\%}{hak-par'in oy yuzdesi, the percentage of votes the Rights and Freedoms Party received}
  \item{TKP_\%}{tkp'nin oy yuzdesi, the percentage of votes the Communist Party of Turkey received}
  \item{TKH_\%}{tkh'nin oy yuzdesi, the percentage of votes the People's Liberation Party received}
  \item{SOL_PARTI_\%}{sol parti'nin oy yuzdesi, the percentage of votes the Left Party received}
  \item{GENCPARTI_\%}{gencparti'nin oy yuzdesi, the percentage of votes the Young Party received}
  \item{MEMLEKET_\%}{memleket partisinin oy yuzdesi, the percentage of votes the Homeland Party received}
  \item{BBP_\%}{bbp'nin oy yuzdesi, the percentage of votes the Great Unity Party received}
  \item{AK_Parti_\%}{ak parti'nin oy yuzdesi, the percentage of votes the Justice and Development Party received}
  \item{YENIDEN_REFAH_\%}{yeniden refah partisinin oy yuzdesi, the percentage of votes the Refah Party received}
  \item{MHP_\%}{mhp'nin oy yuzdesi, the percentage of votes the Nationalist Movement Party received}
  \item{YESIL_SOL_PARTI_\%}{yesil sol partinin oy yuzdesi, the percentage of votes the Green Left Party received}
  \item{TIP_\%}{tip'in oy yuzdesi, the percentage of votes the Labour Party received}
  \item{AB_\%}{ab'nin oy yuzdesi, the percentage of votes the Bloc Party received}
  \item{ANAP_\%}{anap'in oy yuzdesi, the percentage of votes the Motherland Party received}
  \item{YP_\%}{yp'nin oy yuzdesi, the percentage of votes the New Path Party received}
  \item{HKP_\%}{hkp'nin oy yuzdesi, the percentage of votes the People's Liberation Party received}
  \item{MILLI_YOL_\%}{milli yol'un oy yuzdesi, the percentage of votes the National Course Party received}
  \item{VATAN_PARTISI_\%}{vatan partisinin oy yuzdesi, the percentage of votes the Homeland Party received}
  \item{GBP_\%}{gbp'nin oy yuzdesi, the percentage of votes the Independent Turkey Party received}
  \item{CHP_\%}{chp'nin oy yuzdesi, the percentage of votes the Republican People's Party received}
  \item{IYI_Parti_\%}{iyi parti'nin oy yuzdesi, the percentage of votes the Good Party received}
  \item{AP_\%}{ap'nin oy yuzdesi, the percentage of votes the Alternative Party received}
  \item{ZAFER_PARTISI_\%}{zafer partisinin oy yuzdesi, the percentage of votes the Victory Party received}
  \item{SOSYALIST_GUC_BIRLIGI_ITTIFAKI_\%}{sosyalist guc birligi ittifakinin oy yuzdesi, the percentage of votes the Socialist Power Union Alliance received}
  \item{CUMHUR_ITTIFAKI_\%}{cumhur ittifakinin oy yuzdesi, the percentage of votes the People's Alliance received}
  \item{EMEK_VE_OZGURLUK_ITTIFAKI_\%}{emek ve ozgurluk ittifakinin oy yuzdesi, the percentage of votes the Labour and Freedom Alliance received}
  \item{MILLET_ITTIFAKI_\%}{millet ittifakinin oy yuzdesi, the percentage of votes the Nation Alliance received}
  \item{ATA_ITTIFAKI_\%}{ata ittifakinin oy yuzdesi, the percentage of votes the Ancestor Alliance received}
  \item{BAGIMSIZ_TOPLAM_OY_\%}{bagimsiz adaylarin toplam oy yuzdesi, the percentage of votes independent candidates received}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
genel_secim_2023_il
}
\description{
A dataset showing the information for each political party and alliance competing in the general election, their individual vote counts, percentages, and the overall voter statistics by province.
}
\keyword{datasets}
