% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_province_and_party.R
\name{show_province_and_party}
\alias{show_province_and_party}
\title{show_province_and_party}
\usage{
show_province_and_party(province = "Ankara", party = "CHP", percentage = FALSE)
}
\arguments{
\item{province}{The name of the province that users want to look for}

\item{party}{The name of the political party that users want to look for}

\item{percentage}{It is a boolean argument which allows you to print the plot with the percentage or not}
}
\value{
A visual that shows the election results of a specified political party in a given province
}
\description{
A function that displays the election results of a specified political party in a given province, both as percentages and as absolute figures, in the form of a graph.
}
\examples{
\donttest{
show_province_and_party()
}
}
