% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_result_for_turkey.R
\name{show_result_for_turkey}
\alias{show_result_for_turkey}
\title{show_result_for_turkey}
\usage{
show_result_for_turkey(x, percentage = FALSE)
}
\arguments{
\item{x}{the name of the data frame to be visualized. It is supposeed to be the one including name turkey in the file name.}

\item{percentage}{It is a boolean argument which allows you to print the plot with the percentage or not}
}
\value{
A visual that shows the percentages for each parties for the given election result
}
\description{
A function that visualizes election results across Turkey, with the option to display the results either as percentages or as absolute numbers.
}
\examples{
\donttest{
show_result_for_turkey()
}
}
