% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2004_buyuksehir_turkiye}
\alias{yerel_secim_2004_buyuksehir_turkiye}
\title{Information for the results of the local elections in Turkey in 2004}
\format{
A data frame with 21 rows and 6 variables:
\describe{
\item{Parti}{parti, the name of the political party}
\item{Genel_baskani}{genel baskani, the name of the party's leader}
\item{Oy}{oy, the number of votes received by the party}
\item{Yuzde}{yuzde, the percentage of votes received by the party}
\item{Belediye_Baskanligi_Sayisi}{belediye baskanligi sayisi, the number of municipality mayor positions won by the party}
\item{B.Belediye_Baskanligi_Sayisi}{buyuksehir belediye baskanligi sayisi, the number of metropolitan municipality mayor positions won by the party}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2004_buyuksehir_turkiye
}
\description{
A dataset showing the information for the political parties competing in the local elections, their vote counts, percentages, and the number of municipality and metropolitan municipality mayoral positions won.
}
\keyword{datasets}
