% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2009_belediye_turkiye}
\alias{yerel_secim_2009_belediye_turkiye}
\title{Information for the results of the local elections in Turkey in 2009}
\format{
A data frame with 20 rows and 5 variables:
\describe{
  \item{Parti}{parti, the name of the political party}
  \item{Genel_baskani}{genel baskani, the general chairperson of the political party}
  \item{Oy}{oy, the total number of votes received by the party}
  \item{Yuzde}{yuzde, the percentage of votes received by the party}
  \item{Belediye_Baskanligi_Sayisi}{belediye baskanligi sayisi, the number of mayoral positions won by the party}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2009_belediye_turkiye
}
\description{
A dataset showing the information for political parties competing in the local elections, their total votes, vote percentages, and the number of mayoral positions won.
}
\keyword{datasets}
