% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2009_buyuksehir_turkiye}
\alias{yerel_secim_2009_buyuksehir_turkiye}
\title{Information for the results of the metropolitan municipality mayoral elections in Turkey in 2009}
\format{
A data frame with 21 rows and 5 variables:
\describe{
  \item{Parti}{parti, the name of the political party}
  \item{Genel_baskani}{genel baskani, the general chairperson of the political party}
  \item{Oy}{oy, the total number of votes received by the party for metropolitan municipality mayoral positions}
  \item{Yuzde}{yuzde, the percentage of votes received by the party for metropolitan municipality mayoral positions}
  \item{B.Belediye_Baskanligi_Sayisi}{buyuksehir belediye baskanligi sayisi, the number of metropolitan mayoral positions won by the party}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2009_buyuksehir_turkiye
}
\description{
A dataset showing the information for political parties competing in the metropolitan municipality mayoral elections, their total votes, vote percentages, and the number of metropolitan mayoral positions won.
}
\keyword{datasets}
