% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2009_il}
\alias{yerel_secim_2009_il}
\title{Information for the results of the local elections in Turkey by province in 2009, detailing votes for each party}
\format{
A data frame with rows for each province and 44 variables:
\describe{
  \item{Il_Adi}{il adi, the name of the province}
  \item{Kayitli_Secmen_Sayisi}{kayitli secmen sayisi, the number of registered voters in each province}
  \item{Oy_Kullanan_Secmen_Sayisi}{oy kullanan secmen sayisi, the number of voters who voted in each province}
  \item{Gecerli_Oy_Toplami}{gecerli oy toplami, the total number of valid votes in each province}
  \item{ANAVATAN}{anavatan partisinin oy sayisi, the number of votes received by the ANAVATAN party in each province}
  \item{MHP}{mhp'nin oy sayisi, the number of votes received by the Nationalist Movement Party (MHP) in each province}
  \item{LDP}{ldp'nin oy sayisi, the number of votes received by the Liberal Democratic Party (LDP) in each province}
  \item{DSP}{dsp'nin oy sayisi, the number of votes received by the Democratic Left Party (DSP) in each province}
  \item{DTP}{dtp'nin oy sayisi, the number of votes received by the Democratic Society Party (DTP) in each province}
  \item{IP}{ip'nin oy sayisi, the number of votes received by the Independent Turkey Party (IP) in each province}
  \item{CHP}{chp'nin oy sayisi, the number of votes received by the Republican People's Party (CHP) in each province}
  \item{HAK-PAR}{hak-par'in oy sayisi, the number of votes received by the Rights and Freedoms Party (HAK-PAR) in each province}
  \item{BBP}{bbp'nin oy sayisi, the number of votes received by the Great Unity Party (BBP) in each province}
  \item{BDP}{bdp'nin oy sayisi, the number of votes received by the Peace and Democracy Party (BDP) in each province}
  \item{MILLET}{millet partisinin oy sayisi, the number of votes received by the Nation Party (MILLET) in each province}
  \item{ODP}{odp'nin oy sayisi, the number of votes received by the Freedom and Solidarity Party (ODP) in each province}
  \item{TKP}{tkp'nin oy sayisi, the number of votes received by the Communist Party of Turkey (TKP) in each province}
  \item{DP}{dp'nin oy sayisi, the number of votes received by the Democratic Party (DP) in each province}
  \item{SAADET}{saadet partisinin oy sayisi, the number of votes received by the Felicity Party (SAADET) in each province}
  \item{AK_PARTI}{ak parti'nin oy sayisi, the number of votes received by the Justice and Development Party (AK PARTI) in each province}
  \item{BTP}{btp'nin oy sayisi, the number of votes received by the Independent Turkey Party (BTP) in each province}
  \item{EMEP}{emep'in oy sayisi, the number of votes received by the Labour Party (EMEP) in each province}
  \item{HYP}{hyp'nin oy sayisi, the number of votes received by the People's Ascent Party (HYP) in each province}
  \item{BAGIMSIZ_TOPLAM_OY}{bagimsiz adaylarin toplam oy sayisi, the total number of votes received by independent candidates in each province}
  \item{ANAVATAN_\%}{anavatan partisinin oy yuzdesi, the percentage of votes received by the ANAVATAN party in each province}
  \item{MHP_\%}{mhp'nin oy yuzdesi, the percentage of votes received by the Nationalist Movement Party (MHP) in each province}
  \item{LDP_\%}{ldp'nin oy yuzdesi, the percentage of votes received by the Liberal Democratic Party (LDP) in each province}
  \item{DSP_\%}{dsp'nin oy yuzdesi, the percentage of votes received by the Democratic Left Party (DSP) in each province}
  \item{DTP_\%}{dtp'nin oy yuzdesi, the percentage of votes received by the Democratic Society Party (DTP) in each province}
  \item{IP_\%}{ip'nin oy yuzdesi, the percentage of votes received by the Independent Turkey Party (IP) in each province}
  \item{CHP_\%}{chp'nin oy yuzdesi, the percentage of votes received by the Republican People's Party (CHP) in each province}
  \item{HAK-PAR_\%}{hak-par'in oy yuzdesi, the percentage of votes received by the Rights and Freedoms Party (HAK-PAR) in each province}
  \item{BBP_\%}{bbp'nin oy yuzdesi, the percentage of votes received by the Great Unity Party (BBP) in each province}
  \item{BDP_\%}{bdp'nin oy yuzdesi, the percentage of votes received by the Peace and Democracy Party (BDP) in each province}
  \item{MILLET_\%}{millet partisinin oy yuzdesi, the percentage of votes received by the Nation Party (MILLET) in each province}
  \item{ODP_\%}{odp'nin oy yuzdesi, the percentage of votes received by the Freedom and Solidarity Party (ODP) in each province}
  \item{TKP_\%}{tkp'nin oy yuzdesi, the percentage of votes received by the Communist Party of Turkey (TKP) in each province}
  \item{DP_\%}{dp'nin oy yuzdesi, the percentage of votes received by the Democratic Party (DP) in each province}
  \item{SAADET_\%}{saadet partisinin oy yuzdesi, the percentage of votes received by the Felicity Party (SAADET) in each province}
  \item{AK_PARTI_\%}{ak parti'nin oy yuzdesi, the percentage of votes received by the Justice and Development Party (AK PARTI) in each province}
  \item{BTP_\%}{btp'nin oy yuzdesi, the percentage of votes received by the Independent Turkey Party (BTP) in each province}
  \item{EMEP_\%}{emep'in oy yuzdesi, the percentage of votes received by the Labour Party (EMEP) in each province}
  \item{HYP_\%}{hyp'nin oy yuzdesi, the percentage of votes received by the People's Ascent Party (HYP) in each province}
  \item{BAGIMSIZ_TOPLAM_OY_\%}{bagimsiz adaylarin toplam oy yuzdesi, the total percentage of votes received by independent candidates in each province}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2009_il
}
\description{
A dataset showing detailed information for each political party competing in the local elections in every province, including vote counts, percentages, and overall voting statistics.
}
\keyword{datasets}
