% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2014_buyuksehir_turkiye}
\alias{yerel_secim_2014_buyuksehir_turkiye}
\title{Information for the results of the local elections in Turkey in 2014, focusing on municipal and metropolitan mayoral positions}
\format{
A data frame with rows for each political party and 6 variables:
\describe{
\item{Parti}{parti, the name of the political party}
\item{Genel_baskan(lar)}{genel baskan(lar), the general chairperson(s) of the political party}
\item{Oy}{oy, the total number of votes received by the party}
\item{Yuzde}{yuzde, the percentage of votes received by the party}
\item{B.Belediye_Baskanligi_Sayisi}{buyuksehir belediye baskanligi sayisi, the number of metropolitan mayoral positions won by the party}
\item{Belediye_Baskanligi_Sayisi}{belediye baskanligi sayisi, the number of municipal mayoral positions won by the party}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2014_buyuksehir_turkiye
}
\description{
A dataset showing the information for political parties competing in the local elections, their total votes, vote percentages, and the number of municipal and metropolitan mayoral positions won.
}
\keyword{datasets}
