% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2014_il}
\alias{yerel_secim_2014_il}
\title{Information for the results of the local elections in Turkey by province in 2014, detailing votes for each party, including percentages}
\format{
A data frame with rows for 81 rows and 52 variables:
\describe{
\item{Il_Adi}{il adi, the name of the province}
\item{Kayitli_Secmen_Sayisi}{kayitli secmen sayisi, the number of registered voters in each province}
\item{Oy_Kullanan_Secmen_Sayisi}{oy kullanan secmen sayisi, the number of voters who voted in each province}
\item{Gecerli_Oy_Toplami}{gecerli oy toplami, the total number of valid votes in each province}
\item{HUDA_PAR}{huda-par'in oy sayisi, the number of votes the Free Cause Party (HUDA-PAR) received in each province}
\item{HUDA_PAR_\%}{huda-par'in oy yuzdesi, the percentage of votes the Free Cause Party (HUDA-PAR) received in each province}
\item{DSP}{dsp'nin oy sayisi, the number of votes the Democratic Left Party (DSP) received in each province}
\item{DSP_\%}{dsp'nin oy yuzdesi, the percentage of votes the Democratic Left Party (DSP) received in each province}
\item{DYP}{dyp'nin oy sayisi, the number of votes the True Path Party (DYP) received in each province}
\item{DYP_\%}{dyp'nin oy yuzdesi, the percentage of votes the True Path Party (DYP) received in each province}
\item{TURK_PARTI}{turk partisi'nin oy sayisi, the number of votes the Turkish Party received in each province}
\item{TURK_PARTI_\%}{turk partisi'nin oy yuzdesi, the percentage of votes the Turkish Party received in each province}
\item{HKP}{hkp'nin oy sayisi, the number of votes the People's Liberation Party (HKP) received in each province}
\item{HKP_\%}{hkp'nin oy yuzdesi, the percentage of votes the People's Liberation Party (HKP) received in each province}
\item{TKP}{tkp'nin oy sayisi, the number of votes the Communist Party of Turkey (TKP) received in each province}
\item{TKP_\%}{tkp'nin oy yuzdesi, the percentage of votes the Communist Party of Turkey (TKP) received in each province}
\item{BBP}{bbp'nin oy sayisi, the number of votes the Great Unity Party (BBP) received in each province}
\item{BBP_\%}{bbp'nin oy yuzdesi, the percentage of votes the Great Unity Party (BBP) received in each province}
\item{AK_PARTI}{ak parti'nin oy sayisi, the number of votes the Justice and Development Party (AK PARTI) received in each province}
\item{AK_PARTI_\%}{ak parti'nin oy yuzdesi, the percentage of votes the Justice and Development Party (AK PARTI) received in each province}
\item{YURT-P}{yurt-p'nin oy sayisi, the number of votes the Homeland Party (YURT-P) received in each province}
\item{YURT-P_\%}{yurt-p'nin oy yuzdesi, the percentage of votes the Homeland Party (YURT-P) received in each province}
\item{DP}{dp'nin oy sayisi, the number of votes the Democratic Party (DP) received in each province}
\item{DP_\%}{dp'nin oy yuzdesi, the percentage of votes the Democratic Party (DP) received in each province}
\item{EMEP}{emep'in oy sayisi, the number of votes the Labour Party (EMEP) received in each province}
\item{EMEP_\%}{emep'in oy yuzdesi, the percentage of votes the Labour Party (EMEP) received in each province}
\item{ODP}{odp'nin oy sayisi, the number of votes the Freedom and Solidarity Party (ODP) received in each province}
\item{ODP_\%}{odp'nin oy yuzdesi, the percentage of votes the Freedom and Solidarity Party (ODP) received in each province}
\item{MILLET}{millet partisinin oy sayisi, the number of votes the Nation Party (MILLET) received in each province}
\item{MILLET_\%}{millet partisinin oy yuzdesi, the percentage of votes the Nation Party (MILLET) received in each province}
\item{SAADET}{saadet partisinin oy sayisi, the number of votes the Felicity Party (SAADET) received in each province}
\item{SAADET_\%}{saadet partisinin oy yuzdesi, the percentage of votes the Felicity Party (SAADET) received in each province}
\item{HAK-PAR}{hak-par'in oy sayisi, the number of votes the Rights and Freedoms Party (HAK-PAR) received in each province}
\item{HAK-PAR_\%}{hak-par'in oy yuzdesi, the percentage of votes the Rights and Freedoms Party (HAK-PAR) received in each province}
\item{LDP}{ldp'nin oy sayisi, the number of votes the Liberal Democratic Party (LDP) received in each province}
\item{LDP_\%}{ldp'nin oy yuzdesi, the percentage of votes the Liberal Democratic Party (LDP) received in each province}
\item{BDP}{bdp'nin oy sayisi, the number of votes the Peace and Democracy Party (BDP) received in each province}
\item{BDP_\%}{bdp'nin oy yuzdesi, the percentage of votes the Peace and Democracy Party (BDP) received in each province}
\item{BTP}{btp'nin oy sayisi, the number of votes the Independent Turkey Party (BTP) received in each province}
\item{BTP_\%}{btp'nin oy yuzdesi, the percentage of votes the Independent Turkey Party (BTP) received in each province}
\item{IP}{ip'nin oy sayisi, the number of votes the Workers' Party (IP) received in each province}
\item{IP_\%}{ip'nin oy yuzdesi, the percentage of votes the Workers' Party (IP) received in each province}
\item{HDP}{hdp'nin oy sayisi, the number of votes the Peoples' Democratic Party (HDP) received in each province}
\item{HDP_\%}{hdp'nin oy yuzdesi, the percentage of votes the Peoples' Democratic Party (HDP) received in each province}
\item{CHP}{chp'nin oy sayisi, the number of votes the Republican People's Party (CHP) received in each province}
\item{CHP_\%}{chp'nin oy yuzdesi, the percentage of votes the Republican People's Party (CHP) received in each province}
\item{HEPAR}{hepar'in oy sayisi, the number of votes the Rights and Equality Party (HEPAR) received in each province}
\item{HEPAR_\%}{hepar'in oy yuzdesi, the percentage of votes the Rights and Equality Party (HEPAR) received in each province}
\item{MHP}{mhp'nin oy sayisi, the number of votes the Nationalist Movement Party (MHP) received in each province}
\item{MHP_\%}{mhp'nin oy yuzdesi, the percentage of votes the Nationalist Movement Party (MHP) received in each province}
\item{BAGIMSIZ_TOPLAM_OY}{bagimsiz adaylarin toplam oy sayisi, the total number of votes received by independent candidates in each province}
\item{BAGIMSIZ_TOPLAM_OY_\%}{bagimsiz adaylarin toplam oy yuzdesi, the total percentage of votes received by independent candidates in each province}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2014_il
}
\description{
A dataset showing detailed information for each political party competing in the local elections in every province, including vote counts, vote percentages, and overall voting statistics.
}
\keyword{datasets}
