% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2019_buyuksehir_turkiye}
\alias{yerel_secim_2019_buyuksehir_turkiye}
\title{Information for the overall results of the local elections in Turkey in 2019, focusing on total vote percentages and mayoral positions won by parties}
\format{
A data frame with rows for each political party and 6 variables:
\describe{
  \item{Parti}{parti, the name of the political party}
  \item{Genel_baskani}{genel baskani, the general chairperson of the political party}
  \item{Oy}{oy, the total number of votes received by the party nationwide}
  \item{Yuzde}{yuzde, the percentage of total votes received by the party nationwide}
  \item{B.Belediye_Baskanligi_Sayisi}{buyuksehir belediye baskanligi sayisi, the number of metropolitan mayoral positions won by the party nationwide}
  \item{Belediye_Baskanligi_Sayisi}{belediye baskanligi sayisi, the number of municipal mayoral positions won by the party nationwide}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2019_buyuksehir_turkiye
}
\description{
A dataset showing the information for political parties competing in the local elections, their total votes, vote percentages, and the number of municipal and metropolitan mayoral positions won nationwide.
}
\keyword{datasets}
