% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2024_il}
\alias{yerel_secim_2024_il}
\title{Information for the 2024 Local Election Results in Turkey by Province
A dataset showing detailed information for the overall results of the 2024 local elections, including vote counts and percentages for all major political parties across Turkish provinces.}
\format{
A data frame with rows representing various provinces and variables detailing the vote counts and percentages for each party:
\describe{
\item{İl Adı}{il adı, the name of the province.}
\item{Kayıtlı Seçmen Sayısı}{kayıtlı seçmen sayısı, the total number of registered voters.}
\item{Oy Kullanan Seçmen Sayısı}{oy kullanan seçmen sayısı, the total number of voters who cast a ballot.}
\item{Geçerli Oy Toplamı}{geçerli oy toplamı, the total number of valid votes.}
\item{AK PARTİ}{AK PARTİ'nin oy sayısı, the total number of votes for the Justice and Development Party (AK PARTI).}
\item{İYİ PARTİ}{İYİ PARTİ'nin oy sayısı, the total number of votes for the Good Party (İYİ PARTİ).}
\item{SOL PARTİ}{SOL PARTİ'nin oy sayısı, the total number of votes for the Left Party (SOL PARTİ).}
\item{BÜYÜK BİRLİK}{BÜYÜK BİRLİK Partisi'nin oy sayısı, the total number of votes for the Great Unity Party (BÜYÜK BİRLİK).}
\item{MEMLEKET}{MEMLEKET Partisi'nin oy sayısı, the total number of votes for the Homeland Party (MEMLEKET).}
\item{ANAP}{ANAP'ın oy sayısı, the total number of votes for the Motherland Party (ANAP).}
\item{DSP}{DSP'nin oy sayısı, the total number of votes for the Democratic Left Party (DSP).}
\item{YENİDEN REFAH}{YENİDEN REFAH Partisi'nin oy sayısı, the total number of votes for the New Welfare Party (YENİDEN REFAH).}
\item{DEM Parti}{DEM Parti'nin oy sayısı, the total number of votes for the Peoples' Equality and Democracy Party (DEM Parti).}
\item{TKP}{TKP'nin oy sayısı, the total number of votes for the Communist Party of Turkey (TKP).}
\item{ABP}{ABP'nin oy sayısı, the total number of votes for the Anatolian Unity Party (ABP).}
\item{ZAFER PARTİSİ}{ZAFER PARTİSİ'nin oy sayısı, the total number of votes for the Victory Party (ZAFER PARTİSİ).}
\item{HKP}{HKP'nin oy sayısı, the total number of votes for the People's Liberation Party (HKP).}
\item{TKH}{TKH'nin oy sayısı, the total number of votes for the Communist Movement of Turkey (TKH).}
\item{BTP}{BTP'nin oy sayısı, the total number of votes for the Independent Turkey Party (BTP).}
\item{GELECEK PARTİSİ}{GELECEK PARTİSİ'nin oy sayısı, the total number of votes for the Future Party (GELECEK PARTİSİ).}
\item{YTP}{YTP'nin oy sayısı, the total number of votes for the New Turkey Party (YTP).}
\item{CHP}{CHP'nin oy sayısı, the total number of votes for the Republican People's Party (CHP).}
\item{EMEP}{EMEP'in oy sayısı, the total number of votes for the Labour Party (EMEP).}
\item{HÜDA PAR}{HÜDA PAR'ın oy sayısı, the total number of votes for the Free Cause Party (HÜDA PAR).}
\item{HAK-PAR}{HAK-PAR'ın oy sayısı, the total number of votes for the Rights and Freedoms Party (HAK-PAR).}
\item{OCAK}{OCAK Partisi'nin oy sayısı, the total number of votes for the OCAK Party.}
\item{AB PARTİ}{AB PARTİ'nin oy sayısı, the total number of votes for the Justice and Unity Party (AB PARTİ).}
\item{DP}{DP'nin oy sayısı, the total number of votes for the Democratic Party (DP).}
\item{GBP}{GBP'nin oy sayısı, the total number of votes for the Great Homeland Party (GBP).}
\item{MİLLET}{MİLLET Partisi'nin oy sayısı, the total number of votes for the Nation Party (MİLLET).}
\item{MİLLİ YOL}{MİLLİ YOL Partisi'nin oy sayısı, the total number of votes for the National Path Party (MİLLİ YOL).}
\item{AP}{AP'nin oy sayısı, the total number of votes for the Alliance Party (AP).}
\item{ADP}{ADP'nin oy sayısı, the total number of votes for the Anatolian Democratic Party (ADP).}
\item{MHP}{MHP'nin oy sayısı, the total number of votes for the Nationalist Movement Party (MHP).}
\item{TİP}{TİP'in oy sayısı, the total number of votes for the Workers' Party of Turkey (TİP).}
\item{DEVA PARTİSİ}{DEVA PARTİSİ'nin oy sayısı, the total number of votes for the DEVA Party (DEVA PARTİSİ).}
\item{SAADET}{SAADET Partisi'nin oy sayısı, the total number of votes for the Felicity Party (SAADET).}
\item{VATAN PARTİSİ}{VATAN PARTİSİ'nin oy sayısı, the total number of votes for the Homeland Party (VATAN PARTİSİ).}
\item{BAĞIMSIZ TOPLAM OY}{BAĞIMSIZ TOPLAM OY sayısı, the total number of votes received by independent candidates.}
\item{AK PARTİ_\%}{AK PARTİ'nin oy yüzdesi, the percentage of valid votes for AK PARTI.}
\item{İYİ PARTİ_\%}{İYİ PARTİ'nin oy yüzdesi, the percentage of valid votes for İYİ PARTİ.}
\item{SOL PARTİ_\%}{SOL PARTİ'nin oy yüzdesi, the percentage of valid votes for SOL PARTİ.}
\item{BÜYÜK BİRLİK_\%}{BÜYÜK BİRLİK'in oy yüzdesi, the percentage of valid votes for BÜYÜK BİRLİK.}
\item{MEMLEKET_\%}{MEMLEKET'in oy yüzdesi, the percentage of valid votes for MEMLEKET.}
\item{ANAP_\%}{ANAP'ın oy yüzdesi, the percentage of valid votes for ANAP.}
\item{DSP_\%}{DSP'nin oy yüzdesi, the percentage of valid votes for DSP.}
\item{YENİDEN REFAH_\%}{YENİDEN REFAH'ın oy yüzdesi, the percentage of valid votes for YENİDEN REFAH.}
\item{DEM Parti_\%}{DEM Parti'nin oy yüzdesi, the percentage of valid votes for DEM Parti.}
\item{TKP_\%}{TKP'nin oy yüzdesi, the percentage of valid votes for TKP.}
\item{ABP_\%}{ABP'nin oy yüzdesi, the percentage of valid votes for ABP.}
\item{ZAFER PARTİSİ_\%}{ZAFER PARTİSİ'nin oy yüzdesi, the percentage of valid votes for ZAFER PARTİSİ.}
\item{HKP_\%}{HKP'nin oy yüzdesi, the percentage of valid votes for HKP.}
\item{TKH_\%}{TKH'nin oy yüzdesi, the percentage of valid votes for TKH.}
\item{BTP_\%}{BTP'nin oy yüzdesi, the percentage of valid votes for BTP.}
\item{GELECEK PARTİSİ_\%}{GELECEK PARTİSİ'nin oy yüzdesi, the percentage of valid votes for GELECEK PARTİSİ.}
\item{YTP_\%}{YTP'nin oy yüzdesi, the percentage of valid votes for YTP.}
\item{CHP_\%}{CHP'nin oy yüzdesi, the percentage of valid votes for CHP.}
\item{EMEP_\%}{EMEP'in oy yüzdesi, the percentage of valid votes for EMEP.}
\item{HÜDA PAR_\%}{HÜDA PAR'ın oy yüzdesi, the percentage of valid votes for HÜDA PAR.}
\item{HAK-PAR_\%}{HAK-PAR'ın oy yüzdesi, the percentage of valid votes for HAK-PAR.}
\item{OCAK_\%}{OCAK'ın oy yüzdesi, the percentage of valid votes for OCAK.}
\item{AB PARTİ_\%}{AB PARTİ'nin oy yüzdesi, the percentage of valid votes for AB PARTİ.}
\item{DP_\%}{DP'nin oy yüzdesi, the percentage of valid votes for DP.}
\item{GBP_\%}{GBP'nin oy yüzdesi, the percentage of valid votes for GBP.}
\item{MİLLET_\%}{MİLLET'in oy yüzdesi, the percentage of valid votes for MİLLET.}
\item{MİLLİ YOL_\%}{MİLLİ YOL'un oy yüzdesi, the percentage of valid votes for MİLLİ YOL.}
\item{AP_\%}{AP'nin oy yüzdesi, the percentage of valid votes for AP.}
\item{ADP_\%}{ADP'nin oy yüzdesi, the percentage of valid votes for ADP.}
\item{MHP_\%}{MHP'nin oy yüzdesi, the percentage of valid votes for MHP.}
\item{TİP_\%}{TİP'in oy yüzdesi, the percentage of valid votes for TİP.}
\item{DEVA PARTİSİ_\%}{DEVA PARTİSİ'nin oy yüzdesi, the percentage of valid votes for DEVA PARTİSİ.}
\item{SAADET_\%}{SAADET'in oy yüzdesi, the percentage of valid votes for SAADET.}
\item{VATAN PARTİSİ_\%}{VATAN PARTİSİ'nin oy yüzdesi, the percentage of valid votes for VATAN PARTİSİ.}
\item{BAĞIMSIZ TOPLAM OY_\%}{BAĞIMSIZ TOPLAM OY'un oy yüzdesi, the total percentage of valid votes received by independent candidates.}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2024_il
}
\description{
Information for the 2024 Local Election Results in Turkey by Province
A dataset showing detailed information for the overall results of the 2024 local elections, including vote counts and percentages for all major political parties across Turkish provinces.
}
\keyword{datasets}
