% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{yerel_secim_2024_turkiye}
\alias{yerel_secim_2024_turkiye}
\title{Information for the overall results of the 2024 Local Elections in Turkey
A dataset showing the total number of votes and the number of mayoral positions (Metropolitan, District, and Town) won by each political party nationwide in the 2024 Local Elections.}
\format{
A data frame with rows for each political party and 9 variables:
\describe{
\item{Parti}{parti, the standardized name of the political party.}
\item{Genel_baskani}{genel başkanı, the general chairperson/leader of the political party at the time of the election.}
\item{Ittifak}{ittifak, the political alliance the party participated in (if any).}
\item{Belde}{belde, the number of mayoral positions won in towns (Belde) nationwide.}
\item{İlçe}{ilçe, the number of mayoral positions won in districts (İlçe) nationwide.}
\item{Büyükşehir}{büyükşehir, the number of mayoral positions won in metropolitan municipalities (Büyükşehir) nationwide.}
\item{Toplam}{toplam, the total number of mayoral positions won across all three categories (Belde + İlçe + Büyükşehir).}
\item{Oy}{oy, the total number of votes received by the party nationwide.}
\item{Yuzde}{yüzde, the percentage of total votes received by the party nationwide.}
}
}
\source{
{YSK (Supreme Election Council)}
}
\usage{
yerel_secim_2024_turkiye
}
\description{
Information for the overall results of the 2024 Local Elections in Turkey
A dataset showing the total number of votes and the number of mayoral positions (Metropolitan, District, and Town) won by each political party nationwide in the 2024 Local Elections.
}
\keyword{datasets}
