% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_parliament.R
\name{show_parliament}
\alias{show_parliament}
\title{show_result_for_turkey}
\usage{
show_parliament(x)
}
\arguments{
\item{x}{The name of the data frame which is for the general election, so the name should contain genel in its name.}
}
\value{
A visual that shows the parliament for the given election result
}
\description{
A function that visualizes parliament with respect to election results across Turkey
}
\examples{
\donttest{
show_parliament()
}
}
